added scripts for collision and projectile damage using hitboxes
