+++ Assets/Monsters/Purple_Monster_assets/pm+prefab.prefab
@@ -72,7 +72,10 @@ PrefabInstance:
       addedObject: {fileID: 4203737001651982940}
     - targetCorrespondingSourceObject: {fileID: -2253911629886400664, guid: ad6c944c1a8f7ec4ebbde751b3761721, type: 3}
       insertIndex: -1
-      addedObject: {fileID: -2059982632500286825}
+      addedObject: {fileID: 8939992556977218053}
+    - targetCorrespondingSourceObject: {fileID: -2253911629886400664, guid: ad6c944c1a8f7ec4ebbde751b3761721, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 9083472653305041560}
   m_SourcePrefab: {fileID: -8435245712485981826, guid: ad6c944c1a8f7ec4ebbde751b3761721, type: 3}
 --- !u!1 &2372783886298195916 stripped
 GameObject:
@@ -120,7 +123,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   player: {fileID: 0}
   speed: 20
---- !u!114 &-2059982632500286825
+--- !u!114 &8939992556977218053
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -133,3 +136,39 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   damageAmount: 5
+  inHitBox: 0
+--- !u!70 &9083472653305041560
+CapsuleCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2372783886298195916}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_ForceSendLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ForceReceiveLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ContactCaptureLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_CallbackLayers:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_IsTrigger: 0
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: -0.0050000227, y: 0.07533578}
+  m_Size: {x: 0.11268591, y: 0.052925974}
+  m_Direction: 1
