+++ Assets/Scripts/CollissionDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc31692fe80ea1b439d5cc17d19bb516
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
